+++ interop/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [1.0.3](https://github.com/libp2p/js-libp2p/compare/transport-interop-libp2p-main-v1.0.2...transport-interop-libp2p-main-v1.0.3) (2025-09-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/websockets bumped from ^10.0.1 to ^10.0.2
+    * @libp2p/webtransport bumped from ^6.0.2 to ^6.0.3
+
+## [1.0.2](https://github.com/libp2p/js-libp2p/compare/transport-interop-libp2p-main-v1.0.1...transport-interop-libp2p-main-v1.0.2) (2025-09-25)
+
+
+### Bug Fixes
+
+* bump noise version ([#3306](https://github.com/libp2p/js-libp2p/issues/3306)) ([71e8ee1](https://github.com/libp2p/js-libp2p/commit/71e8ee1632fdeaff5d6c33a38ae6df02ea69f579))
+* use @chainsafe/libp2p-noise and @chainsafe/libp2p-yamux ([#3308](https://github.com/libp2p/js-libp2p/issues/3308)) ([425a42c](https://github.com/libp2p/js-libp2p/commit/425a42cddac5aac4d0ac822295cc4c4817dcdc95))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/mplex bumped from ^12.0.1 to ^12.0.2
+    * @libp2p/webrtc bumped from ^6.0.1 to ^6.0.2
+    * @libp2p/webtransport bumped from ^6.0.1 to ^6.0.2
+    * libp2p bumped from ^3.0.1 to ^3.0.2
+
 ## [1.0.1](https://github.com/libp2p/js-libp2p/compare/transport-interop-libp2p-main-v1.0.0...transport-interop-libp2p-main-v1.0.1) (2025-09-24)